+++ src/CustomRouteSwitch/context.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const CustomRouteSwitchContext = createContext([]);
+
+export default CustomRouteSwitchContext;