+++ src/CustomRouteSwitch/index.js
@@ -1,4 +1,4 @@
 import CustomRouteSwitch from './customRouteSwitch';
 
 export default CustomRouteSwitch;
-export { findValidChildren, validators, createValidator } from './validators';
+export { validators, createValidator } from './validators';