+++ src/index.js
@@ -1,4 +1,6 @@
 import CustomRouteSwitch from './CustomRouteSwitch';
+import CustomRouteCase from './CustomRouteCase';
 
 export default CustomRouteSwitch;
-export { findValidChildren, validators, createValidator } from './CustomRouteSwitch';
+export { validators, createValidator } from './CustomRouteSwitch';
+export { CustomRouteSwitch, CustomRouteCase };